Fixed game structed and added abstract class SDLGameObject

+++ TextureManager.cpp
@@ -22,7 +22,7 @@ bool TextureManager::load(std::string fileName, std::string id, SDL_Renderer* te
 	return false;
 };
 
-void TextureManager::draw(std::string id, int x_pos, int y_pos, int width, int height, SDL_Renderer* textureRenderer, SDL_FlipMode flip)
+void TextureManager::draw(std::string id, float x_pos, float y_pos, float width, float height, SDL_Renderer* textureRenderer, SDL_FlipMode flip)
 {
 	SDL_FRect sourceRect = {};
 	SDL_FRect targetRect = {};
@@ -38,7 +38,7 @@ void TextureManager::draw(std::string id, int x_pos, int y_pos, int width, int h
 	SDL_RenderTextureRotated(textureRenderer, textureMap[id], &sourceRect, &targetRect, 0.0, NULL, flip);
 };
 
-void TextureManager::drawFrame(std::string id, int x_pos, int y_pos, int width, int height, int currentRow, int currentFrame, SDL_Renderer* textureRenderer, SDL_FlipMode flip)
+void TextureManager::drawFrame(std::string id, float x_pos, float y_pos, float width, float height, int currentRow, int currentFrame, SDL_Renderer* textureRenderer, SDL_FlipMode flip)
 {
 	SDL_FRect sourceRect = {};
 	SDL_FRect targetRect = {};
