+++ TextureManager.h
@@ -15,8 +15,8 @@ public:
 	}
 
 	bool load(std::string fileName, std::string id, SDL_Renderer* textureRenderer);
-	void draw(std::string id, int x_pos, int y_pos, int width, int height, SDL_Renderer* textureRenderer, SDL_FlipMode flip = SDL_FLIP_NONE);
-	void drawFrame(std::string id, int x_pos, int y_pos, int width, int height, int currentRow, int currentFrame, SDL_Renderer* textureRenderer, SDL_FlipMode flip = SDL_FLIP_NONE);
+	void draw(std::string id, float x_pos, float y_pos, float width, float height, SDL_Renderer* textureRenderer, SDL_FlipMode flip = SDL_FLIP_NONE);
+	void drawFrame(std::string id, float x_pos, float y_pos, float width, float height, int currentRow, int currentFrame, SDL_Renderer* textureRenderer, SDL_FlipMode flip = SDL_FLIP_NONE);
 	std::map<std::string, SDL_Texture*> textureMap;
 	static TextureManager* theInstance;
 private:
